pyproject.toml
[project]
name = "ai-xspan-test"
version = "0.1.0"
description = "Agentic RAG on Archival Research Papers"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-community>=0.3",
    "openai>=1.0",
    "chromadb>=0.5,<0.6",
    "onnxruntime==1.19.2",
    "rank-bm25>=0.2",
    "pymupdf>=1.24",
    "sentence-transformers>=3.0,<4.0",
    "torch==2.2.2",
    "numpy>=1.26,<2.0",
    "python-dotenv>=1.0",
    "gradio>=4.0",
    "tiktoken>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
